requirements.txt
alabaster==0.7.13
arrow==1.2.3
astroid==2.15.6
asttokens==2.4.0
atomicwrites==1.4.1
attrs==23.1.0
autopep8==2.0.4
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.9.1
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
cftime==1.6.2
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
comm==0.1.4
contourpy==1.1.0
cookiecutter==2.3.0
cryptography==41.0.3
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
docstring-to-markdown==0.12
docutils==0.20.1
executing==1.2.0
fastjsonschema==2.18.0
flake8==6.0.0
fonttools==4.42.1
gssapi==1.8.3
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.25.2
ipython==8.12.2
ipython-genutils==0.2.0
isort==5.12.0
jaraco.classes==3.3.0
jedi==0.18.2
jeepney==0.8.0
jellyfish==1.0.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
keyring==24.2.0
kiwisolver==1.4.5
krb5==0.5.1
lazy-object-proxy==1.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.1
more-itertools==10.1.0
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
netCDF4==1.6.4
numpy==1.24.4
numpydoc==1.5.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.16.1
pylint==2.17.5
pylint-venv==3.0.2
pyls-spyder==0.4.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pysmb==1.2.9.1
pyspnego==0.9.2
python-dateutil==2.8.2
python-lsp-black==1.3.0
python-lsp-jsonrpc==1.1.1
python-lsp-server==1.7.4
python-slugify==8.0.1
pytoolconfig==1.2.5
pytz==2023.3.post1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.1
QDarkStyle==3.1
qstylizer==0.2.2
QtAwesome==1.2.3
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rope==1.9.0
rpds-py==0.10.3
Rtree==1.0.1
SecretStorage==3.3.3
six==1.16.0
smbprotocol==1.11.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.5
spyder-kernels==2.4.4
sshtunnel==0.4.0
stack-data==0.6.2
text-unidecode==1.3
textdistance==4.5.0
three-merge==0.1.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
tunnel==1.0.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
whatthepatch==1.0.5
wrapt==1.15.0
wurlitzer==3.0.3
xarray==2023.1.0
yapf==0.40.1
zipp==3.16.2
